requirements.txt
fastapi
python-multipart
python-jose[cryptography]
uvicorn
